[proxima-core] #35 use executor for running observers
