+++ cassandra/src/main/java/cz/o2/proxima/storage/cassandra/DateToLongConverter.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2017 O2 Czech Republic, a.s.
+ * Copyright 2017-2018 O2 Czech Republic, a.s.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package cz.o2.proxima.storage.cassandra;
 
 import cz.seznam.euphoria.shaded.guava.com.google.common.base.Strings;
@@ -30,7 +29,7 @@ public class DateToLongConverter implements StringConverter<Date> {
   private static final Date MIN = new Date(Long.MIN_VALUE);
 
   @Override
-  public String toString(Date what) {
+  public String asString(Date what) {
     return String.valueOf(what.getTime());
   }
