+++ cassandra/src/main/java/cz/o2/proxima/storage/cassandra/StringConverter.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2017 O2 Czech Republic, a.s.
+ * Copyright 2017-2018 O2 Czech Republic, a.s.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package cz.o2.proxima.storage.cassandra;
 
 import java.io.Serializable;
@@ -33,7 +32,7 @@ public interface StringConverter<T> extends Serializable {
     final String MIN = "";
 
     @Override
-    public String toString(String what) {
+    public String asString(String what) {
       return (String) what;
     }
 
@@ -57,7 +56,7 @@ public interface StringConverter<T> extends Serializable {
   /**
    * Convert type to string.
    */
-  String toString(T what);
+  String asString(T what);
 
 
   /**
