+++ core/src/main/java/cz/o2/proxima/repository/Context.java
@@ -1,0 +1,61 @@
+/**
+ * Copyright 2017-2018 O2 Czech Republic, a.s.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package cz.o2.proxima.repository;
+
+import cz.seznam.euphoria.core.client.functional.VoidFunction;
+import java.io.Serializable;
+import java.util.concurrent.ScheduledExecutorService;
+
+/**
+ * Context created in local instance that can be distributed over wire.
+ */
+public class Context implements Serializable {
+
+  /**
+   * Executor associated with all asynchronous operations.
+   */
+  private final VoidFunction<ScheduledExecutorService> executorFactory;
+
+  /**
+   * Initialization marker. After deserialization this will be `false`.
+   */
+  private transient boolean initialized = false;
+
+  /**
+   * Materialized executor.
+   */
+  private transient ScheduledExecutorService service;
+
+  protected Context(VoidFunction<ScheduledExecutorService> executorFactory) {
+    this.executorFactory = executorFactory;
+  }
+
+  /**
+   * Get executor for asynchronous tasks.
+   */
+  public ScheduledExecutorService getExecutorService() {
+    initialize();
+    return service;
+  }
+
+  private synchronized void initialize() {
+    if (!initialized) {
+      service = executorFactory.apply();
+      initialized = true;
+    }
+  }
+
+}
