+++ core/src/main/java/cz/o2/proxima/storage/DataAccessor.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2017 O2 Czech Republic, a.s.
+ * Copyright 2017-2018 O2 Czech Republic, a.s.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package cz.o2.proxima.storage;
 
+import cz.o2.proxima.repository.Context;
 import cz.o2.proxima.storage.batch.BatchLogObservable;
 import cz.o2.proxima.storage.commitlog.CommitLogReader;
 import cz.o2.proxima.storage.randomaccess.RandomAccessReader;
@@ -31,35 +31,35 @@ public interface DataAccessor extends Serializable {
   /**
    * Retrieve writer (if applicable).
    */
-  default Optional<AttributeWriterBase> getWriter() {
+  default Optional<AttributeWriterBase> getWriter(Context context) {
     return Optional.empty();
   }
 
   /**
    * Retrieve commit log reader (if applicable).
    */
-  default Optional<CommitLogReader> getCommitLogReader() {
+  default Optional<CommitLogReader> getCommitLogReader(Context context) {
     return Optional.empty();
   }
 
   /**
    * Retrieve random access reader.
    */
-  default Optional<RandomAccessReader> getRandomAccessReader() {
+  default Optional<RandomAccessReader> getRandomAccessReader(Context context) {
     return Optional.empty();
   }
 
   /**
    * Retrieve batch log observable.
    */
-  default Optional<BatchLogObservable> getBatchLogObservable() {
+  default Optional<BatchLogObservable> getBatchLogObservable(Context context) {
     return Optional.empty();
   }
 
   /**
    * Retrieve partitioned view of the data.
    */
-  default Optional<PartitionedView> getPartitionedView() {
+  default Optional<PartitionedView> getPartitionedView(Context context) {
     return Optional.empty();
   }
