+++ core/src/test/java/cz/o2/proxima/storage/InMemBulkStorage.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2017 O2 Czech Republic, a.s.
+ * Copyright 2017-2018 O2 Czech Republic, a.s.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package cz.o2.proxima.storage;
 
+import cz.o2.proxima.repository.Context;
 import cz.o2.proxima.repository.EntityDescriptor;
 import java.net.URI;
 import java.util.Collections;
@@ -68,7 +68,7 @@ public class InMemBulkStorage extends StorageDescriptor {
     }
 
     @Override
-    public Optional<AttributeWriterBase> getWriter() {
+    public Optional<AttributeWriterBase> getWriter(Context context) {
       return Optional.of(new Writer(entityDesc, uri));
     }
