+++ kafka/src/main/java/cz/o2/proxima/storage/kafka/KafkaLogReader.java
@@ -16,43 +16,20 @@
 package cz.o2.proxima.storage.kafka;
 
 import cz.o2.proxima.repository.AttributeDescriptor;
-import cz.o2.proxima.repository.EntityDescriptor;
-import cz.o2.proxima.storage.AbstractOnlineAttributeWriter;
-import cz.o2.proxima.storage.AttributeWriterBase;
-import cz.o2.proxima.storage.CommitCallback;
-import cz.o2.proxima.storage.DataAccessor;
+import cz.o2.proxima.repository.Context;
+import cz.o2.proxima.storage.AbstractStorage;
 import cz.o2.proxima.storage.Partition;
-import cz.o2.proxima.storage.StreamElement;
 import cz.o2.proxima.storage.commitlog.BulkLogObserver;
 import cz.o2.proxima.storage.commitlog.Cancellable;
 import cz.o2.proxima.storage.commitlog.CommitLogReader;
 import cz.o2.proxima.storage.commitlog.LogObserver;
 import cz.o2.proxima.storage.commitlog.LogObserverBase;
-import cz.o2.proxima.storage.kafka.partitioner.KeyPartitioner;
-import cz.o2.proxima.util.Classpath;
 import cz.o2.proxima.view.PartitionedLogObserver;
 import cz.o2.proxima.view.PartitionedView;
 import cz.o2.proxima.view.input.DataSourceUtils;
 import cz.seznam.euphoria.core.client.dataset.Dataset;
 import cz.seznam.euphoria.core.client.flow.Flow;
 import cz.seznam.euphoria.core.client.operator.MapElements;
-import cz.seznam.euphoria.shaded.guava.com.google.common.base.Strings;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.clients.consumer.OffsetAndMetadata;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.common.PartitionInfo;
-import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.serialization.Serdes;
-
-import javax.annotation.Nullable;
-import java.net.URI;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
@@ -60,234 +37,49 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
 import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiFunction;
-import java.util.function.Consumer;
 import java.util.stream.Collectors;
+import javax.annotation.Nullable;
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.clients.consumer.ConsumerRecords;
+import org.apache.kafka.clients.consumer.KafkaConsumer;
+import org.apache.kafka.clients.consumer.OffsetAndMetadata;
+import org.apache.kafka.common.PartitionInfo;
+import org.apache.kafka.common.TopicPartition;
 
 /**
- * Kafka writer and commit log using {@code KafkaProducer}.
+ * A {@link CommitLogReader} implementation for Kafka.
  */
 @Slf4j
-public class KafkaCommitLog extends AbstractOnlineAttributeWriter
-    implements CommitLogReader, DataAccessor, PartitionedView {
-
-  /**
-   * Data read from a kafka partition.
-   */
-  private static final class KafkaStreamElement extends StreamElement {
-
-    @Getter
-    private final int partition;
-    @Getter
-    private final long offset;
-
-    KafkaStreamElement(
-        EntityDescriptor entityDesc,
-        AttributeDescriptor attributeDesc,
-        String uuid, String key, String attribute,
-        long stamp, byte[] value, int partition, long offset) {
-
-      super(entityDesc, attributeDesc, uuid, key, attribute, stamp, value);
-      this.partition = partition;
-      this.offset = offset;
-    }
-
-    @Override
-    public String toString() {
-      return "KafkaStreamElement(entityDesc=" + getEntityDescriptor()
-          + ", attributeDesc=" + getAttributeDescriptor()
-          + ", uuid=" + getUuid()
-          + ", key=" + getKey()
-          + ", attribute=" + getAttribute()
-          + ", stamp=" + getStamp()
-          + ", value.length=" + (getValue() == null ? 0 : getValue().length)
-          + ", partition=" + partition
-          + ", offset=" + offset
-          + ")";
-    }
-
-  }
-
-  /**
-   * Consumer of stream elements.
-   * The callback may or might not be called depending on the consuming mode
-   * (bulk or online).
-   */
-  static interface ElementConsumer {
-    void consumeWithConfirm(
-        @Nullable StreamElement element,
-        TopicPartition tp, long offset,
-        Consumer<Throwable> errorHandler);
-  }
-
-  static interface TopicPartitionCommitter {
-    void commit(TopicPartition tp, long offset);
-  }
-
-  static final class OnlineConsumer implements ElementConsumer {
-    final LogObserver observer;
-    final TopicPartitionCommitter committer;
-    OnlineConsumer(LogObserver observer, TopicPartitionCommitter committer) {
-      this.observer = observer;
-      this.committer = committer;
-    }
-
-    @Override
-    public void consumeWithConfirm(
-        @Nullable StreamElement element,
-        TopicPartition tp, long offset,
-        Consumer<Throwable> errorHandler) {
-
-      if (element != null) {
-        observer.onNext(element, tp::partition, (succ, exc) -> {
-          if (succ) {
-            committer.commit(tp, offset);
-          } else {
-            errorHandler.accept(exc);
-          }
-        });
-      } else {
-        committer.commit(tp, offset);
-      }
-    }
-  }
-
-  static final class BulkConsumer implements ElementConsumer {
-    final BulkLogObserver observer;
-    final TopicPartitionCommitter committer;
-    BulkConsumer(BulkLogObserver observer, TopicPartitionCommitter committer) {
-      this.observer = observer;
-      this.committer = committer;
-    }
-
-    @Override
-    public void consumeWithConfirm(
-        @Nullable StreamElement element,
-        TopicPartition tp, long offset,
-        Consumer<Throwable> errorHandler) {
-
-      if (element != null) {
-        observer.onNext(element, tp::partition, (succ, exc) -> {
-          if (succ) {
-            committer.commit(tp, offset);
-          } else {
-            errorHandler.accept(exc);
-          }
-        });
-      }
-    }
-  }
-
-  /** A poll interval in milliseconds. */
-  public static final String POLL_INTERVAL_CFG = "poll.interval";
-  /** Partitioner class for entity key-attribute pair. */
-  public static final String PARTITIONER_CLASS = "partitioner";
-
-  public static final String WRITER_CONFIG_PREFIX = "kafka.";
-  private static final int PRODUCE_CONFIG_PREFIX_LENGTH = WRITER_CONFIG_PREFIX.length();
+public class KafkaLogReader extends AbstractStorage
+    implements CommitLogReader, PartitionedView {
 
   @Getter
+  final KafkaAccessor accessor;
+  private final Context context;
+  private final AtomicBoolean shutdown = new AtomicBoolean();
+  private final long consumerPollInterval;
   private final String topic;
 
-  private final Map<String, Object> cfg;
-  private final AtomicBoolean shutdown = new AtomicBoolean();
-
-  private KafkaProducer<String, byte[]> producer;
-  private long consumerPollInterval = 100;
-
-  Partitioner partitioner = new KeyPartitioner();
-
-  public KafkaCommitLog(
-      EntityDescriptor entity,
-      URI uri,
-      Map<String, Object> cfg) {
-
-    super(entity, uri);
-
-    if (uri.getPath().length() <= 1) {
-      throw new IllegalArgumentException("Specify topic by path in URI");
-    }
-    if (Strings.isNullOrEmpty(uri.getAuthority())) {
-      throw new IllegalArgumentException("Specify brokers by authority in URI");
-    }
-
-    this.cfg = cfg;
-    this.topic = Utils.topic(uri);
-    configure(cfg);
+  KafkaLogReader(KafkaAccessor accessor, Context context) {
+    super(accessor.getEntityDescriptor(), accessor.getURI());
+    this.accessor = accessor;
+    this.context = context;
+    this.consumerPollInterval = accessor.getConsumerPollInterval();
+    this.topic = accessor.getTopic();
   }
-
-
-  private void configure(Map<String, Object> cfg) {
-    this.consumerPollInterval = Optional.ofNullable(cfg.get(POLL_INTERVAL_CFG))
-        .map(v -> Long.valueOf(v.toString()))
-        .orElse(consumerPollInterval);
-
-    this.partitioner = Optional.ofNullable((String) cfg.get(PARTITIONER_CLASS))
-        .map(cls -> Classpath.findClass(cls, Partitioner.class))
-        .map(cls -> {
-          try {
-            return cls.newInstance();
-          } catch (InstantiationException | IllegalAccessException ex) {
-            throw new RuntimeException(ex);
-          }
-        })
-        .orElse(this.partitioner);
-
-    log.info(
-        "Using consumerPollInterval {} and partitionerClass {} for URI {}",
-        consumerPollInterval, partitioner.getClass(), getURI());
-  }
-
-
-  @Override
-  public List<Partition> getPartitions() {
-    final List<PartitionInfo> partitions;
-    if (this.producer == null) {
-      try (KafkaConsumer<String, byte[]> consumer = createConsumer()) {
-        partitions = consumer.partitionsFor(topic);
-      }
-    } else {
-      partitions = this.producer.partitionsFor(topic);
-    }
-    return partitions.stream().map(pi ->
-      new Partition() {
-        @Override
-        public int getId() {
-          return pi.partition();
-        }
-      }
-    ).collect(Collectors.toList());
-  }
-
-
-  @Override
-  @SuppressWarnings("unchecked")
-  public void write(StreamElement data, CommitCallback callback) {
-    try {
-      if (this.producer == null) {
-        this.producer = createProducer();
-      }
-      int partition = (partitioner.getPartitionId(
-          data.getKey(), data.getAttribute(), data.getValue()) & Integer.MAX_VALUE)
-          % producer.partitionsFor(topic).size();
-      producer.send(
-          new ProducerRecord(topic, partition, data.getStamp(), data.getKey()
-              + "#" + data.getAttribute(), data.getValue()),
-              (metadata, exception) -> callback.commit(exception == null, exception));
-    } catch (Exception ex) {
-      log.warn("Failed to write ingest {}", data, ex);
-      callback.commit(false, ex);
-    }
-  }
-
 
   /**
    * Subscribe observer by name to the commit log.
@@ -369,6 +161,22 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
         .output();
   }
 
+  @Override
+  public List<Partition> getPartitions() {
+    final List<PartitionInfo> partitions;
+    try (KafkaConsumer<String, byte[]> consumer = createConsumer()) {
+      partitions = consumer.partitionsFor(topic);
+    }
+    return partitions.stream().map(pi ->
+      new Partition() {
+        @Override
+        public int getId() {
+          return pi.partition();
+        }
+      }
+    ).collect(Collectors.toList());
+  }
+
   protected Cancellable observePartitions(
       @Nullable String name,
       @Nullable Collection<Partition> partitions,
@@ -380,8 +188,10 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
     // wait until the consumer is really created
     CountDownLatch latch = new CountDownLatch(1);
 
+    ExecutorService executor = context.getExecutorService();
+
     // start new thread that will fill our observer
-    Thread consumer = new Thread(() -> {
+    Future<?> submit = executor.submit(() -> {
       try {
         try (KafkaConsumer<String, byte[]> kafkaConsumer = createConsumer(
             name, partitions, listener, position)) {
@@ -403,9 +213,6 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
         observer.onError(thwbl);
       }
     });
-    consumer.setDaemon(true);
-    consumer.setName("consumer-" + name);
-    consumer.start();
 
     try {
       log.debug("Waiting for the consumer {} to be created and run", name);
@@ -416,7 +223,7 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
     }
 
     return () -> {
-      consumer.interrupt();
+      submit.cancel(true);
     };
   }
 
@@ -520,7 +327,7 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
         return null;
     };
 
-    OnlineConsumer onlineConsumer = new OnlineConsumer(observer, (tp, offset) -> {
+    KafkaAccessor.OnlineConsumer onlineConsumer = new KafkaAccessor.OnlineConsumer(observer, (tp, offset) -> {
         if (named) {
           offsetCommitter.confirm(tp, offset);
         }
@@ -553,7 +360,7 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
       return null;
     };
 
-    BulkConsumer bulkConsumer = new BulkConsumer(observer, (tp, offset) -> {
+    KafkaAccessor.BulkConsumer bulkConsumer = new KafkaAccessor.BulkConsumer(observer, (tp, offset) -> {
       if (named) {
         Map<TopicPartition, OffsetAndMetadata> copy;
         synchronized (uncommittedMap) {
@@ -581,7 +388,7 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
       boolean named,
       boolean stopAtCurrent,
       BiFunction<TopicPartition, ConsumerRecord<String, byte[]>, Void> preWrite,
-      ElementConsumer consumer,
+      KafkaAccessor.ElementConsumer consumer,
       Map<TopicPartition, OffsetAndMetadata> commitMap,
       LogObserverBase baseObserver) {
 
@@ -664,34 +471,6 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
   }
 
 
-  @SuppressWarnings("unchecked")
-  private KafkaProducer<String, byte[]> createProducer() {
-    Properties props = createProps();
-    props.put(ProducerConfig.ACKS_CONFIG, "all");
-    props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, getURI().getAuthority());
-    props.put(ProducerConfig.BATCH_SIZE_CONFIG, 16384);
-    return new KafkaProducer<>(
-        props, Serdes.String().serializer(), Serdes.ByteArray().serializer());
-  }
-
-
-  @SuppressWarnings("unchecked")
-  private Properties createProps() {
-    Properties props = new Properties();
-    for (Map.Entry<String, Object> e : cfg.entrySet()) {
-      if (e.getKey().startsWith(WRITER_CONFIG_PREFIX)) {
-        props.put(e.getKey().substring(PRODUCE_CONFIG_PREFIX_LENGTH),
-            e.getValue().toString());
-      }
-    }
-    return props;
-  }
-
-  /** Create kafka consumer with specific rebalance listener. */
-  public KafkaConsumerFactory createConsumerFactory() {
-    return new KafkaConsumerFactory(getURI(), createProps());
-  }
-
   private KafkaConsumer<String, byte[]> createConsumer() {
     return createConsumer(
         "dummy-consumer",
@@ -706,7 +485,7 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
       @Nullable ConsumerRebalanceListener listener,
       Position position) {
 
-    KafkaConsumerFactory factory = createConsumerFactory();
+    KafkaConsumerFactory factory = accessor.createConsumerFactory();
     final KafkaConsumer<String, byte[]> consumer;
 
     if (name != null) {
@@ -734,28 +513,7 @@ public class KafkaCommitLog extends AbstractOnlineAttributeWriter
 
   @Override
   public void close() {
-    if (this.producer != null) {
-      this.producer.close();
-      this.producer = null;
-    }
     this.shutdown.set(true);
   }
-
-  @Override
-  public Optional<AttributeWriterBase> getWriter() {
-    return Optional.of(this);
-  }
-
-  @Override
-  public Optional<CommitLogReader> getCommitLogReader() {
-    return Optional.of(this);
-  }
-
-  @Override
-  public Optional<PartitionedView> getPartitionedView() {
-    return Optional.of(this);
-  }
-
-
 
 }
